feat: property修改

+++ TuneLab.SDK.Base/Property/PropertyNumber_V1.cs
@@ -20,27 +20,27 @@ public class PropertyNumber_V1 : IPrimitiveValue_V1
     public static implicit operator nuint(PropertyNumber_V1 property) => (nuint)property.mValue;
     public static implicit operator float(PropertyNumber_V1 property) => (float)property.mValue;
     public static implicit operator double(PropertyNumber_V1 property) => (double)property.mValue;
-    public static implicit operator decimal(PropertyNumber_V1 property) => property.mValue;
+    public static implicit operator decimal(PropertyNumber_V1 property) => (decimal)property.mValue;
 
-    public static implicit operator PropertyNumber_V1(sbyte value) => new(value);
-    public static implicit operator PropertyNumber_V1(byte value) => new(value);
-    public static implicit operator PropertyNumber_V1(short value) => new(value);
-    public static implicit operator PropertyNumber_V1(ushort value) => new(value);
-    public static implicit operator PropertyNumber_V1(int value) => new(value);
-    public static implicit operator PropertyNumber_V1(uint value) => new(value);
-    public static implicit operator PropertyNumber_V1(long value) => new(value);
-    public static implicit operator PropertyNumber_V1(ulong value) => new(value);
-    public static implicit operator PropertyNumber_V1(nint value) => new(value);
-    public static implicit operator PropertyNumber_V1(nuint value) => new(value);
-    public static implicit operator PropertyNumber_V1(float value) => new((decimal)value);
-    public static implicit operator PropertyNumber_V1(double value) => new((decimal)value);
-    public static implicit operator PropertyNumber_V1(decimal value) => new(value);
+    public static implicit operator PropertyNumber_V1(sbyte value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(byte value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(short value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(ushort value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(int value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(uint value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(long value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(ulong value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(nint value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(nuint value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(float value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(double value) => new((double)value);
+    public static implicit operator PropertyNumber_V1(decimal value) => new((double)value);
 
-    public PropertyNumber_V1(decimal number) { mValue = number; }
+    public PropertyNumber_V1(double number) { mValue = number; }
 
     public override string ToString() => mValue.ToString();
 
     bool IEquatable<IPropertyValue_V1>.Equals(IPropertyValue_V1? other) => other is PropertyNumber_V1 property && property.mValue == mValue;
 
-    readonly decimal mValue;
+    readonly double mValue;
 }
